Split 2d and 3d

+++ crates/bevy_tiles/src/commands/chunk_batch.rs
@@ -4,9 +4,9 @@ use super::{insert_chunk_batch, take_chunk_batch_despawn_tiles};
 
 pub struct SpawnChunkBatch<F, B, IC, const N: usize = 2>
 where
-    F: Fn([isize; N]) -> B + Send + 'static,
+    F: Fn([i32; N]) -> B + Send + 'static,
     B: Bundle + Send + 'static,
-    IC: IntoIterator<Item = [isize; N]> + Send + 'static,
+    IC: IntoIterator<Item = [i32; N]> + Send + 'static,
 {
     pub map_id: Entity,
     pub chunk_cs: IC,
@@ -15,12 +15,12 @@ where
 
 impl<F, B, IC, const N: usize> Command for SpawnChunkBatch<F, B, IC, N>
 where
-    F: Fn([isize; N]) -> B + Send + 'static,
+    F: Fn([i32; N]) -> B + Send + 'static,
     B: Bundle + Send + 'static,
-    IC: IntoIterator<Item = [isize; N]> + Send + 'static,
+    IC: IntoIterator<Item = [i32; N]> + Send + 'static,
 {
     fn apply(self, world: &mut World) {
-        let (chunk_cs, bundles): (Vec<[isize; N]>, Vec<B>) = self
+        let (chunk_cs, bundles): (Vec<[i32; N]>, Vec<B>) = self
             .chunk_cs
             .into_iter()
             .map(|coord| (coord, (self.bundle_f)(coord)))
@@ -29,7 +29,7 @@ where
         let chunks = chunk_cs
             .into_iter()
             .zip(world.spawn_batch(bundles))
-            .collect::<Vec<([isize; N], Entity)>>();
+            .collect::<Vec<([i32; N], Entity)>>();
 
         insert_chunk_batch::<N>(world, self.map_id, chunks);
     }
@@ -37,7 +37,7 @@ where
 
 pub struct DespawnChunkBatch<IC, const N: usize = 2>
 where
-    IC: IntoIterator<Item = [isize; N]> + Send + 'static,
+    IC: IntoIterator<Item = [i32; N]> + Send + 'static,
 {
     pub map_id: Entity,
     pub chunk_cs: IC,
@@ -45,7 +45,7 @@ where
 
 impl<IC, const N: usize> Command for DespawnChunkBatch<IC, N>
 where
-    IC: IntoIterator<Item = [isize; N]> + Send + 'static,
+    IC: IntoIterator<Item = [i32; N]> + Send + 'static,
 {
     fn apply(self, world: &mut World) {
         for (_, tile_id) in take_chunk_batch_despawn_tiles::<N>(world, self.map_id, self.chunk_cs) {
